Added command group for attendance commands. Verification group has verification commands only.
