+++ classes/bot-guild.js
@@ -482,13 +482,8 @@ class BotGuild {
         /** @type {CommandoGuild} */
         let guild = await client.guilds.fetch(this._id);
 
-        guild.setCommandEnabled('verify', this.verification.isEnabled);
-        guild.setCommandEnabled('add-members', this.verification.isEnabled);
-        guild.setCommandEnabled('manual-verify', this.verification.isEnabled);
-        guild.setCommandEnabled('check-member', this.verification.isEnabled | this.attendance.isEnabled);
-        guild.setCommandEnabled('start-attend', this.attendance.isEnabled);
-        guild.setCommandEnabled('attend',  this.attendance.isEnabled);
-        guild.setGroupEnabled('verification', this.attendance.isEnabled || this.verification.isEnabled);
+        guild.setGroupEnabled('verification', this.verification.isEnabled);
+        guild.setGroupEnabled('attendance', this.attendance.isEnabled);
 
         guild.setGroupEnabled('stamps', this.stamps.isEnabled);
