+++ commands/attendance/start-attend.js
@@ -17,7 +17,7 @@ class StartAttend extends PermissionCommand {
     constructor(client) {
         super(client, {
             name: 'start-attend',
-            group: 'a_utility',
+            group: 'attendance',
             memberName: 'initiate attend process',
             description: 'identifies/makes a channel to be used for !attend and notifies people',
             guildOnly: true,
